+++ lib/flutter_map_tile_caching.dart
@@ -1,6 +1,15 @@
 // Copyright © Luka S (JaffaKetchup) under GPL-v3
 // A full license can be found at .\LICENSE
 
+/// A plugin for flutter_map providing advanced caching functionality, with
+/// ability to download map regions for offline use. Also includes useful
+/// prebuilt widgets.
+///
+/// * [GitHub Repository](https://github.com/JaffaKetchup/flutter_map_tile_caching)
+/// * [pub.dev Package](https://pub.dev/packages/flutter_map_tile_caching)
+///
+/// * [Documentation Site](https://fmtc.jaffaketchup.dev/)
+/// * [Full API Reference](https://pub.dev/documentation/flutter_map_tile_caching/latest/flutter_map_tile_caching/flutter_map_tile_caching-library.html)
 library flutter_map_tile_caching;
 
 import 'dart:async';
@@ -13,10 +22,15 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_background/flutter_background.dart';
+import 'package:flutter_background/flutter_background.dart'
+    as flutter_background;
 import 'package:flutter_local_notifications/flutter_local_notifications.dart';
+import 'package:flutter_local_notifications/flutter_local_notifications.dart'
+    as flutter_local_notifications;
 import 'package:flutter_map/flutter_map.dart';
 import 'package:flutter_map/plugin_api.dart';
 import 'package:http/http.dart' as http;
+import 'package:isar/isar.dart' as isar;
 import 'package:isar/isar.dart';
 import 'package:latlong2/latlong.dart';
 import 'package:meta/meta.dart';
@@ -27,7 +41,7 @@ import 'package:queue/queue.dart';
 import 'package:share_plus/share_plus.dart';
 import 'package:stream_transform/stream_transform.dart';
 
-import 'src/bulk_download/download_progress.dart';
+import 'src/bulk_download/bulk_tile_writer.dart';
 import 'src/bulk_download/downloader.dart';
 import 'src/bulk_download/internal_timing_progress_management.dart';
 import 'src/bulk_download/tile_loops.dart';
@@ -38,24 +52,17 @@ import 'src/db/defs/store.dart';
 import 'src/db/defs/tile.dart';
 import 'src/db/registry.dart';
 import 'src/db/tools.dart';
-import 'src/misc/enums.dart';
 import 'src/misc/exts.dart';
-import 'src/providers/tile_provider.dart';
-import 'src/settings/fmtc_settings.dart';
+import 'src/providers/image_provider.dart';
 import 'src/settings/tile_provider_settings.dart';
 
-export 'package:flutter_background/flutter_background.dart'
-    show AndroidResource;
-export 'package:flutter_local_notifications/flutter_local_notifications.dart'
-    show AndroidNotificationDetails;
-
-export 'src/bulk_download/download_progress.dart';
 export 'src/misc/background_download_widget.dart';
-export 'src/misc/enums.dart';
-export 'src/settings/fmtc_settings.dart';
 export 'src/settings/tile_provider_settings.dart';
 
+part 'src/bulk_download/download_progress.dart';
 part 'src/fmtc.dart';
+part 'src/misc/exports.dart';
+part 'src/providers/tile_provider.dart';
 part 'src/regions/base_region.dart';
 part 'src/regions/circle.dart';
 part 'src/regions/downloadable_region.dart';
@@ -68,6 +75,7 @@ part 'src/root/manage.dart';
 part 'src/root/migrator.dart';
 part 'src/root/recovery.dart';
 part 'src/root/statistics.dart';
+part 'src/settings/fmtc_settings.dart';
 part 'src/store/directory.dart';
 part 'src/store/download.dart';
 part 'src/store/export.dart';