+++ lib/src/settings/tile_provider_settings.dart
@@ -3,9 +3,24 @@
 
 import 'package:isar/isar.dart';
 
-import '../misc/enums.dart';
+import '../../flutter_map_tile_caching.dart';
 import '../providers/image_provider.dart';
-import '../providers/tile_provider.dart';
+
+/// Multiple behaviors dictating how browse caching should be carried out
+///
+/// Check documentation on each value for more information.
+enum CacheBehavior {
+  /// Only get tiles from the local cache
+  ///
+  /// Useful for applications with dedicated 'Offline Mode'.
+  cacheOnly,
+
+  /// Get tiles from the local cache, going on the Internet to update the cached tile if it has expired (`cachedValidDuration` has passed)
+  cacheFirst,
+
+  /// Get tiles from the Internet and update the cache for every tile
+  onlineFirst,
+}
 
 /// Settings for an [FMTCTileProvider]
 class FMTCTileProviderSettings {