+++ CHANGELOG.md
@@ -13,11 +13,22 @@ Many thanks to my sponsors, no matter how much or how little they donated. Spons
 
 # Changelog
 
-## [7.0.0] - 2022/12/XX
+## [7.0.0] - 2023/01/XX
 
-* Performance improvements (migrated to Isar)
-* Multiple bug fixes
+* Migrated to Isar database
+* Major performance improvements, thanks to Isar
+* Added buffering to bulk tile downloading
 * Added method to catch tile retrieval errors
+* Removed v4 -> v5 migrator & added v6 -> v7 migrator
+* Removed some synchronous methods from structure management
+* Removed 'fmtc_advanced' import file
+
+Plus the usual:
+
+* Minor performance improvements
+* Bug fixes
+* Dependency updates
+* Documentation improvements
 
 ## [6.2.0] - 2022/10/25